Refactorización de base de datos.
1. Modificaciones de Base de Datos (Migraciones)
Usuarios: Cambié identificacion a BigInteger (seguro contra desbordamientos de datos de cédulas largas). La columna ya se llamaba password, así que ahí estaba segura.
Cursos: Añadí la columna codigo, y renombré estado a activo (booleano).
Horarios:
Cambié hora_inicio y hora_fin a Strings fijos de 5 caracteres ("HH:MM").
Añadí la columna salon.
Cumpliendo tu To-Do List: Dividí formalmente los cupos en cupo_maximo_estudiante, cupo_disponible_estudiante, cupo_maximo_tercero y cupo_disponible_tercero.
2. Actualización de Lógica y Controladores
InscripcionController: Refactorizado para que, al momento en que un usuario se inscriba, el sistema verifique si es un estudiante o un tercero, y reste (o sume al cancelar) el cupo de la columna correspondiente.
CursosController: Creé una lógica dinámica para que al enviarle los horarios al Frontend (student.blade.php), el controlador identifique automáticamente quién solicita la información y le pase los cupos correctos como cupo_disponible. Esto evitó que tuviéramos que reescribir todo el JavaScript de la vista de los estudiantes.
3. Vistas y Modelos
Actualicé la vista administrativa (adminCursos.blade.php) para que al crear un nuevo horario, el admin pueda escribir independientemente cuántos cupos desea abrir para estudiantes y cuántos para terceros, además de poder asignar un salón.
Actualicé todos los arrays $fillable de los modelos Eloquent.

+++ database/migrations/2025_12_27_213649_crear_tabla_usuarios.php
@@ -16,7 +16,7 @@ return new class extends Migration
 
             $table->id();
             $table->string('nombre_apellido');
-            $table->integer('identificacion')->unique();
+            $table->bigInteger('identificacion')->unique();
             $table->string('correo')->unique();
             $table->string('password');
             $table->string('telefono')->nullable();
